NHOMF2-C401/vinfast_route_planner/requirements.txt
streamlit
pytest
openai
python-dotenv
folium
streamlit-folium
pytest